[refactor] YAGNI・DRY・SOLID原則に基づくコード品質改善
- YAGNI原則の適用:
  - 未使用機能の削除 (generateRandomPort, isPrivilegedPort等)
  - 過剰な詳細情報の簡素化 (PortValidationResult, FoundationError)

- DRY原則の強化:
  - プロジェクト関連型の統一 (新規project.ts作成)
  - 重複する型定義の整理

- SOLID原則の改善:
  - 責任分離: port.tsから関数をutils/portUtils.tsに移動
  - 型定義と実装ロジックの明確な分離

- テスト戦略の最適化:
  - 型定義テストの削除 (TypeScriptで保証済み)
  - ビジネスロジックテストのみに集中 (19→8テスト)
  - テスト実行時間短縮 (270ms→205ms)

すべてのテストが通り、ビルドも成功しています。
