+++ packages/shared/src/index.ts
@@ -9,14 +9,14 @@ export type {
   FoundationError,
   PortConfigurationValidationResult,
 } from './types/port';
+export { PORT_RANGE } from './types/port';
+
+// ポート関連ユーティリティ
 export {
-  PORT_RANGE,
   validatePortConfiguration,
   createPortConfiguration,
   isValidPortNumber,
-  isPrivilegedPort,
-  generateRandomPort,
-} from './types/port';
+} from './utils/portUtils';
 
 // ヘルスチェック関連型定義
 export type {
@@ -25,11 +25,13 @@ export type {
   HealthResponse,
   ServerInfo,
   ServerStatus,
-  ProjectInfo,
 } from './types/health';
 
 // API関連型定義
-export type { ApiErrorType, ApiError, ApiResponse, ProjectResponse } from './types/api';
+export type { ApiErrorType, ApiError, ApiResponse } from './types/api';
+
+// プロジェクト関連型定義
+export type { ProjectInfo, ProjectResponse } from './types/project';
 
 // 既存の型定義（後方互換性のため保持）
 export interface FileTreeNode {
