+++ packages/shared/src/types/health.ts
@@ -77,20 +77,3 @@ export interface ServerStatus {
   /** バックエンドサーバー情報 */
   readonly backend: ServerInfo;
 }
-
-/**
- * プロジェクト情報
- *
- * 現在のプロジェクトの基本情報と.kiroディレクトリの存在状況を表します。
- * kiro-lensが動作するプロジェクトのコンテキスト情報を提供します。
- */
-export interface ProjectInfo {
-  /** プロジェクト名（通常はディレクトリ名） */
-  readonly name: string;
-  /** プロジェクトの絶対パス */
-  readonly path: string;
-  /** .kiroディレクトリが存在するかどうか */
-  readonly hasKiroDirectory: boolean;
-  /** .kiroディレクトリの絶対パス（存在する場合のみ） */
-  readonly kiroPath?: string;
-}
